move transformer layer creation up a level to leave room for new TrajEncoders
